# Speedup xml escaping and unescaping