While profiling a server that uses ltx heavily, I noticed that XML escaping/unescaping took quite a lot of the processing time.  I looked for some optimizations and I've found something that yields quite a good improvement.  

Instead of using `.replace` multiple times with different regular expressions, combining the regular expression into a single `.replace` and using a replace function yields much higher throughput for replacement.

When I run the benchmarks, the serialize benchmark runs about 2x faster and the parse benchmark runs 33% faster.

I think the reason this works so well, is that escaping is the uncommon case.  Likely a string that required a lot of escaping or unescaping might be a bit slower, but overall this is a nice performance boost.