@sonnyp 2016-03-30
For reference:

### before
```
sonny@kirby ~/P/n/ltx> npm run benchmark

> ltx@2.2.1 benchmark /home/sonny/Projects/node-xmpp/ltx
> node benchmarks/

suite ltx
tag with template literal x 42,130 ops/sec ±1.46% (86 runs sampled)
tag with direct call x 42,848 ops/sec ±1.97% (87 runs sampled)
serialize and parse x 55,789 ops/sec ±1.29% (91 runs sampled)
parse x 69,463 ops/sec ±1.83% (90 runs sampled)
serialize x 416,451 ops/sec ±1.16% (92 runs sampled)
createElement (jsx) x 446,952 ops/sec ±0.38% (94 runs sampled)
clone x 1,153,505 ops/sec ±0.78% (94 runs sampled)
Element x 1,439,831 ops/sec ±0.68% (92 runs sampled)
Fastest is Element


suite backends parse
Saxjs x 27,825 ops/sec ±1.35% (88 runs sampled)
NodeXML x 14,401 ops/sec ±2.68% (80 runs sampled)
Libxmljs x 29,655 ops/sec ±2.14% (72 runs sampled)
Expat x 50,661 ops/sec ±1.32% (86 runs sampled)
Ltx x 71,840 ops/sec ±1.55% (92 runs sampled)
Fastest is Ltx


suite backends write
Saxjs x 41,556 ops/sec ±0.48% (93 runs sampled)
NodeXML x 48,339 ops/sec ±1.02% (93 runs sampled)
Libxmljs x 75,791 ops/sec ±0.75% (92 runs sampled)
Expat x 118,057 ops/sec ±0.54% (94 runs sampled)
Ltx x 146,653 ops/sec ±0.53% (93 runs sampled)
Fastest is Ltx
```

### after

```
sonny@kirby ~/P/n/ltx> npm run benchmark

> ltx@2.2.1 benchmark /home/sonny/Projects/node-xmpp/ltx
> node benchmarks/

suite ltx
tag with template literal x 57,320 ops/sec ±1.82% (89 runs sampled)
tag with direct call x 60,269 ops/sec ±1.39% (90 runs sampled)
serialize and parse x 73,194 ops/sec ±1.45% (87 runs sampled)
parse x 90,419 ops/sec ±1.58% (93 runs sampled)
serialize x 795,145 ops/sec ±0.05% (95 runs sampled)
createElement (jsx) x 459,494 ops/sec ±0.67% (90 runs sampled)
clone x 1,158,027 ops/sec ±0.59% (96 runs sampled)
Element x 1,469,994 ops/sec ±0.57% (95 runs sampled)
Fastest is Element


suite backends parse
Saxjs x 25,730 ops/sec ±1.84% (87 runs sampled)
NodeXML x 14,994 ops/sec ±2.74% (79 runs sampled)
Libxmljs x 29,649 ops/sec ±1.99% (69 runs sampled)
Expat x 52,577 ops/sec ±1.40% (86 runs sampled)
Ltx x 90,686 ops/sec ±1.64% (90 runs sampled)
Fastest is Ltx


suite backends write
Saxjs x 39,447 ops/sec ±0.78% (89 runs sampled)
NodeXML x 55,986 ops/sec ±0.22% (96 runs sampled)
Libxmljs x 78,542 ops/sec ±0.82% (95 runs sampled)
Expat x 123,103 ops/sec ±0.62% (94 runs sampled)
Ltx x 227,891 ops/sec ±1.08% (89 runs sampled)
Fastest is Ltx

```