@sonnyp 2016-03-30
with node-expat benchmark

### before

```
sonny@kirby ~/P/n/node-expat> npm run benchmark

> node-expat@2.3.13 benchmark /home/sonny/Projects/node-xmpp/node-expat
> node ./benchmark.js

sax x 58,442 ops/sec ±9.73% (54 runs sampled)
node-xml x 102,854 ops/sec ±1.67% (63 runs sampled)
libxmljs x 182,154 ops/sec ±1.93% (63 runs sampled)
node-expat x 235,905 ops/sec ±2.28% (64 runs sampled)
ltx x 300,891 ops/sec ±2.74% (67 runs sampled)
Fastest is ltx
```

### after
```
sonny@kirby ~/P/n/node-expat> npm run benchmark

> node-expat@2.3.13 benchmark /home/sonny/Projects/node-xmpp/node-expat
> node ./benchmark.js

sax x 63,526 ops/sec ±8.58% (57 runs sampled)
node-xml x 102,202 ops/sec ±6.61% (64 runs sampled)
libxmljs x 202,078 ops/sec ±4.89% (66 runs sampled)
node-expat x 243,984 ops/sec ±3.54% (64 runs sampled)
ltx x 461,491 ops/sec ±2.22% (63 runs sampled)
Fastest is ltx
```

which makes ltx parser twice faster than the fastest Node.js XML parser