@sonnyp 2016-03-30
@mdlavin thanks this is a very nice speedup

Out of curiosity, what is your use case?

Removing the abstraction between ltx and parsers might (benchmark needed) also increase op/s significantly.