@mdlavin 2016-03-30
@sonnyp I noticed this issue because I had created a benchmark that removed networking and only dealt with parsing and stringifying stanzas so I could see how my system performed separately from network calls.

In my particular case, I was generating a lot of stanzas in response to joining a large MUC.